requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
exceptiongroup==1.3.1
fastapi==0.123.8
h11==0.16.0
idna==3.11
# Editable install with no version control (iop-sdk-python==1.1.0)
-e /home/ec2-user/iop-sdk
pydantic==2.12.5
pydantic_core==2.41.5
python-dotenv==1.2.1
requests==2.32.5
starlette==0.49.3
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.38.0
